+++ src/app/modules/global-admin/global-admin-organizations/organizations-dialogs/create-organization-dialog/create-organization-dialog.component.ts
@@ -13,9 +13,9 @@ import {
 } from 'src/app/shared/models/organization/organization-odata.model';
 import { cvrValidator } from 'src/app/shared/validators/cvr.validator';
 import { OrganizationActions } from 'src/app/store/organization/actions';
-import { GlobalAdminOrganizationsDialogBaseComponent } from '../global-admin-organizations-dialog-base.component';
 import { OrganizationsDialogComponentStore } from '../organizations-dialog.component-store';
 import { ShallowOptionType } from 'src/app/shared/models/options/option-type.model';
+import { GlobalAdminOrganizationsDialogBaseComponent } from '../global-admin-organizations-dialog-base.component';
 
 @Component({
   selector: 'app-create-organization-dialog',
