+++ src/app/modules/global-admin/global-admin-organizations/organizations-dialogs/global-admin-organizations-dialog-base.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit } from "@angular/core";
-import { OrganizationsDialogComponentStore } from "./organizations-dialog.component-store";
+import { Component, OnInit } from '@angular/core';
+import { OrganizationsDialogComponentStore } from './organizations-dialog.component-store';
 
 @Component({
   template: '',
@@ -9,10 +9,9 @@ export class GlobalAdminOrganizationsDialogBaseComponent implements OnInit {
   protected countryCodes$ = this.componentStore.countryCodes$;
   protected loading$ = this.componentStore.loading$;
 
-  constructor(protected componentStore: OrganizationsDialogComponentStore
-  ) { }
+  constructor(protected componentStore: OrganizationsDialogComponentStore) {}
 
   ngOnInit() {
     this.componentStore.getCountryCodes();
-    }
+  }
 }
